# contracts/compatibility/baseline/openapi/public-v1.yaml
openapi: 3.1.0
info:
    title: Seev public user API
    version: 1.0.0
    description: Stable v1 user contract for auth and gateway-owned product journeys.
servers:
    - {url: 'https://api.example.invalid', description: Gateway public edge}
    - {url: 'https://auth.example.invalid', description: Auth service public listener}
...
paths:
...
    /api/v1/topup:
        post:
            operationId: gatewayCreateTopupV1
            tags: [payments]
            x-see-owner: gateway
            x-see-behavior-owner: payin
            x-see-audience: public
            security: [{bearerAuth: []}]
            requestBody: {$ref: {required: true, content: {application/json: {schema: {$ref: {type: object, additionalProperties: true, description: Operation-specific request schemas tighten this shape; this shared fallback is not a permission to accept unknown fields.}}}}}}
            responses: {'201': {$ref: {description: Successful JSON response., headers: {X-Request-ID: {description: Correlation identifier echoed by the server., schema: {type: string, minLength: 1, maxLength: 128}}}, content: {application/json: {schema: {$ref: {type: object, required: [success, data], additionalProperties: false, properties: {success: {const: true}, data: {}, meta: {type: object, additionalProperties: true}}}}}}}}, '400': {$ref: {description: 'Invalid request, including unknown fields.', content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '401': {$ref: {description: Authentication is required or invalid., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '409': {$ref: {description: Idempotency or state conflict., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '422': {$ref: {description: 'Invalid request, including unknown fields.', content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}}
    /api/v1/topup/{id}:
        get:
            operationId: gatewayGetTopupV1
            tags: [payments]
            x-see-owner: gateway
            x-see-behavior-owner: payin
            x-see-audience: public
            security: [{bearerAuth: []}]
            parameters: [{$ref: {name: id, in: path, required: true, schema: {$ref: {type: string, format: uuid, examples: ['00000000-0000-7000-8000-000000000001']}}}}]
            responses: {'200': {$ref: {description: Successful JSON response., headers: {X-Request-ID: {description: Correlation identifier echoed by the server., schema: {type: string, minLength: 1, maxLength: 128}}}, content: {application/json: {schema: {$ref: {type: object, required: [success, data], additionalProperties: false, properties: {success: {const: true}, data: {}, meta: {type: object, additionalProperties: true}}}}}}}}, '401': {$ref: {description: Authentication is required or invalid., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '404': {$ref: {description: Resource was not found., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}}
    /api/v1/payout:
        post:
            operationId: gatewayCreatePayoutV1
            tags: [payments]
            x-see-owner: gateway
            x-see-behavior-owner: payout
            x-see-audience: public
            security: [{bearerAuth: []}]
            requestBody: {$ref: {required: true, content: {application/json: {schema: {$ref: {type: object, additionalProperties: true, description: Operation-specific request schemas tighten this shape; this shared fallback is not a permission to accept unknown fields.}}}}}}
            responses: {'201': {$ref: {description: Successful JSON response., headers: {X-Request-ID: {description: Correlation identifier echoed by the server., schema: {type: string, minLength: 1, maxLength: 128}}}, content: {application/json: {schema: {$ref: {type: object, required: [success, data], additionalProperties: false, properties: {success: {const: true}, data: {}, meta: {type: object, additionalProperties: true}}}}}}}}, '400': {$ref: {description: 'Invalid request, including unknown fields.', content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '401': {$ref: {description: Authentication is required or invalid., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '409': {$ref: {description: Idempotency or state conflict., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '422': {$ref: {description: 'Invalid request, including unknown fields.', content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}}
    /api/v1/payout/{id}:
        get:
            operationId: gatewayGetPayoutV1
            tags: [payments]
            x-see-owner: gateway
            x-see-behavior-owner: payout
            x-see-audience: public
            security: [{bearerAuth: []}]
            parameters: [{$ref: {name: id, in: path, required: true, schema: {$ref: {type: string, format: uuid, examples: ['00000000-0000-7000-8000-000000000001']}}}}]
            responses: {'200': {$ref: {description: Successful JSON response., headers: {X-Request-ID: {description: Correlation identifier echoed by the server., schema: {type: string, minLength: 1, maxLength: 128}}}, content: {application/json: {schema: {$ref: {type: object, required: [success, data], additionalProperties: false, properties: {success: {const: true}, data: {}, meta: {type: object, additionalProperties: true}}}}}}}}, '401': {$ref: {description: Authentication is required or invalid., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '404': {$ref: {description: Resource was not found., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}}
    /api/v1/notifications:
        get:
            operationId: gatewayListNotificationsV1
            tags: [notifications]
            x-see-owner: gateway
            x-see-behavior-owner: notify
            x-see-audience: public
            security: [{bearerAuth: []}]
            parameters: [{name: cursor, in: query, schema: {type: string, maxLength: 512}}, {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}]
            responses: {'200': {$ref: {description: Successful JSON response., headers: {X-Request-ID: {description: Correlation identifier echoed by the server., schema: {type: string, minLength: 1, maxLength: 128}}}, content: {application/json: {schema: {$ref: {type: object, required: [success, data], additionalProperties: false, properties: {success: {const: true}, data: {}, meta: {type: object, additionalProperties: true}}}}}}}}, '401': {$ref: {description: Authentication is required or invalid., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}}
    /api/v1/notifications/{id}/read:
        post:
            operationId: gatewayMarkNotificationReadV1
            tags: [notifications]
            x-see-owner: gateway
            x-see-behavior-owner: notify
            x-see-audience: public
            security: [{bearerAuth: []}]
            parameters: [{$ref: {name: id, in: path, required: true, schema: {$ref: {type: string, format: uuid, examples: ['00000000-0000-7000-8000-000000000001']}}}}]
            responses: {'204': {$ref: {description: Operation completed without a response body.}}, '401': {$ref: {description: Authentication is required or invalid., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '404': {$ref: {description: Resource was not found., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}}
    /api/v1/ledger/transactions:
        post:
            operationId: gatewayPostLedgerTransactionV1
            tags: [ledger]
            x-see-owner: gateway
            x-see-behavior-owner: ledger
            x-see-audience: public
            security: [{bearerAuth: []}]
            requestBody: {$ref: {required: true, content: {application/json: {schema: {$ref: {type: object, additionalProperties: true, description: Operation-specific request schemas tighten this shape; this shared fallback is not a permission to accept unknown fields.}}}}}}
            responses: {'201': {$ref: {description: Successful JSON response., headers: {X-Request-ID: {description: Correlation identifier echoed by the server., schema: {type: string, minLength: 1, maxLength: 128}}}, content: {application/json: {schema: {$ref: {type: object, required: [success, data], additionalProperties: false, properties: {success: {const: true}, data: {}, meta: {type: object, additionalProperties: true}}}}}}}}, '400': {$ref: {description: 'Invalid request, including unknown fields.', content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '401': {$ref: {description: Authentication is required or invalid., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '403': {$ref: {description: Authenticated principal lacks permission., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '409': {$ref: {description: Idempotency or state conflict., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '422': {$ref: {description: 'Invalid request, including unknown fields.', content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '429': {$ref: {description: Rate limit exceeded., headers: {Retry-After: {schema: {type: integer, minimum: 1}}}, content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}, '503': {$ref: {description: A dependency is temporarily unavailable., content: {application/json: {schema: {$ref: {type: object, required: [success, error], additionalProperties: false, properties: {success: {const: false}, error: {$ref: {type: object, required: [code, message], additionalProperties: false, properties: {code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}, message: {type: string, minLength: 1, maxLength: 512}, details: {type: object, additionalProperties: true}}}}}}}}}}}}
...
components:
    securitySchemes:
        bearerAuth: {$ref: {type: http, scheme: bearer, bearerFormat: JWT}}